feat: update ulysse

+++ src/utils.mjs
@@ -2,52 +2,28 @@ import fs from 'fs';
 import os from 'os';
 import uti from 'util';
 import { exec } from 'child_process';
+import { Service } from 'node-linux';
 
-const DEFAULT_CONFIG_PATH = `${os.homedir()}/.config/ulysse/config.json`;
-const DNSMASQ_CONFIG_PATH = process.env.DNSMASQ_CONFIG_PATH || '/etc/dnsmasq.conf';
+const DEFAULT_CONFIG_PATH = `${process.env.SUDO_USER ? `/home/${process.env.SUDO_USER}` : os.homedir()}/.config/ulysse/config.json`;
+
 const HOSTS_CONFIG_PATH = process.env.HOSTS_CONFIG_PATH || '/etc/hosts';
 
 const params = process.argv.slice(2);
+
 const configPath = params.includes('--config') ? params[params.indexOf('--config') + 1] : DEFAULT_CONFIG_PATH;
 
 export const config = JSON.parse(await fs.readFileSync(configPath, 'utf8'));
 
-export const blockAllHosts = async (whitelist) => {
-    const server = config.server.replace('https://', '');
+export const sleep = async (ms) => new Promise((r) => { setTimeout(r, ms); });
 
-    const lines = [
-        'domain-needed',
-        'bogus-priv',
-        'no-resolv',
-        'server=9.9.9.9',
-        'server=/toggl.com/#',
-        `server=/${server}/#`,
-        'address=/#/0.0.0.0',
-        'address=/#/::',
-        '',
-    ].join('\n');
+export const blockHosts = async () => {
+    const { blocklist } = config;
 
-    fs.writeFileSync(DNSMASQ_CONFIG_PATH, lines, 'utf8');
-    fs.appendFileSync(DNSMASQ_CONFIG_PATH, whitelist.map((host) => `server=/${host}/#`).join('\n'), 'utf8');
+    if (!blocklist.length) return;
 
-    await exec('systemctl restart dnsmasq');
-};
+    console.log('Block hosts...');
 
-export const blockHosts = async (blocklist) => {
     fs.writeFileSync(HOSTS_CONFIG_PATH, blocklist.map((host) => `127.0.0.1 ${host} www.${host}`).join('\n'), 'utf8');
-};
-
-export const unBlockAllHosts = async () => {
-    const lines = [
-        'domain-needed',
-        'bogus-priv',
-        'no-resolv',
-        'server=9.9.9.9',
-    ].join('\n');
-
-    fs.writeFileSync(DNSMASQ_CONFIG_PATH, lines, 'utf8');
-
-    await exec('systemctl restart dnsmasq');
 };
 
 export const unBlockHosts = async () => {
@@ -62,20 +38,29 @@ export const unBlockHosts = async () => {
     fs.writeFileSync(HOSTS_CONFIG_PATH, lines, 'utf8');
 };
 
-export const blockApps = async (apps) => {
+export const blockApps = async () => {
+    const { apps } = config;
+
+    if (!apps.length) return;
+
+    console.log('Block apps...');
+
     for await (const app of apps) {
         await exec(`chmod -x /usr/bin/${app}`);
         await exec(`pkill -f -9 ${app}`);
     }
 };
 
-export const unBlockApps = async (apps) => {
+export const unBlockApps = async () => {
+    const { apps } = config;
+
     for await (const app of apps) {
         await exec(`chmod +x /usr/bin/${app}`);
     }
 };
 
 export const blockRoot = async () => {
+    console.log('Block root...');
     await exec('chmod -x /usr/bin/sudo');
     await exec('chmod -x /usr/bin/su');
     await exec('chmod -x /usr/bin/sudoedit');
@@ -89,12 +74,37 @@ export const unBlockRoot = async () => {
     await exec('chmod +x /usr/bin/suexec');
 };
 
-export const clearBrowser = async () => {
-    await exec('pkill -TERM chromium');
-    await exec('rm -rf ~/.config/chromium/Default/History');
-    await exec('rm -rf ~/.config/chromium/Default/Local Storage');
-    await exec('rm -rf ~/.cache/chromium');
-    await exec('chromium');
+export const checkRoot = async () => {
+    const execAsync = uti.promisify(exec);
+    const { stdout } = await execAsync('whoami').catch(() => false);
+
+    if (stdout.trim() !== 'root') {
+        console.error('You must be root to run this command');
+        process.exit(1);
+    }
+};
+
+export const installDaemon = async () => {
+    await checkRoot();
+
+    const originalConsoleLog = console.log;
+    console.log = () => {};
+
+    const svc = new Service({
+        name: 'ulysse',
+        description: 'Ulysse',
+        script: `./src/daemon.mjs --config ${configPath}`,
+    });
+
+    if (!svc.exists()) {
+        originalConsoleLog('Install daemon...');
+        await svc.install();
+        await sleep(1000);
+    }
+
+    await svc.start();
+
+    console.log = originalConsoleLog;
 };
 
 export const checkDaemon = async () => {
@@ -106,5 +116,3 @@ export const checkDaemon = async () => {
         process.exit(1);
     }
 };
-
-export const sleep = async (ms) => new Promise((r) => { setTimeout(r, ms); });
